Moving contract hooks into their own folder
