+++ hooks/calls/useRepay.js
@@ -4,8 +4,8 @@ import { parseUnits } from "@ethersproject/units";
 import { useWeb3React } from "@web3-react/core";
 import LENDING_MARKET_ABI from "constants/abis/lendingMarket.json";
 import useCurrentToken from "hooks/useCurrentToken";
-import useMarketRegistryContract from "hooks/useMarketRegistryContract";
-import useTokenContract from "hooks/useTokenContract";
+import useMarketRegistryContract from "hooks/contracts/useMarketRegistryContract";
+import useTokenContract from "hooks/contracts/useTokenContract";
 import { useCallback } from "react";
 
 export default function useRepay() {
