+++ hooks/calls/useWithdrawRewards.js
@@ -1,6 +1,6 @@
 import { useWeb3React } from "@web3-react/core";
 import useCurrentToken from "hooks/useCurrentToken";
-import useUnionContract from "hooks/useUnionContract";
+import useUnionContract from "hooks/contracts/useUnionContract";
 import { useCallback } from "react";
 
 export default function useWithdrawRewards() {
