+++ hooks/contracts/useAssetContract.js
@@ -1,7 +1,7 @@
 import { useWeb3React } from "@web3-react/core";
 import ABI from "constants/abis/assetManager.json";
 import { ASSET_MANAGER_ADDRESSES } from "constants/variables";
-import useContract from "./useContract";
+import useContract from "../useContract";
 
 export default function useAssetContract() {
   const { chainId } = useWeb3React();
