+++ hooks/contracts/useControllerContract.js
@@ -1,7 +1,7 @@
 import { useWeb3React } from "@web3-react/core";
 import ABI from "constants/abis/controller.json";
 import { CONTROLLER_ADDRESSES } from "constants/variables";
-import useContract from "./useContract";
+import useContract from "../useContract";
 
 export default function useControllerContract() {
   const { chainId } = useWeb3React();
