+++ hooks/contracts/useMarketRegistryContract.js
@@ -1,7 +1,7 @@
 import { useWeb3React } from "@web3-react/core";
 import ABI from "constants/abis/marketRegistry.json";
 import { MARKET_REGISTRY_ADDRESSES } from "constants/variables";
-import useContract from "./useContract";
+import useContract from "../useContract";
 
 export default function useMarketRegistryContract() {
   const { chainId } = useWeb3React();
