+++ hooks/contracts/useMemberContract.js
@@ -1,7 +1,7 @@
 import { useWeb3React } from "@web3-react/core";
 import ABI from "constants/abis/memberManager.json";
 import { MEMBER_MANAGER_ADDRESSES } from "constants/variables";
-import useContract from "./useContract";
+import useContract from "../useContract";
 
 export default function useMemberContract() {
   const { chainId } = useWeb3React();
