+++ hooks/contracts/useTokenContract.js
@@ -1,5 +1,5 @@
 import ABI from "constants/abis/erc20Detailed.json";
-import useContract from "./useContract";
+import useContract from "../useContract";
 
 export default function useTokenContract(tokenAddress) {
   return useContract(tokenAddress, ABI);
