+++ hooks/contracts/useUnionContract.js
@@ -1,7 +1,7 @@
 import { useWeb3React } from "@web3-react/core";
 import ABI from "constants/abis/unionToken.json";
 import { UNION_TOKEN_ADDRESSES } from "constants/variables";
-import useContract from "./useContract";
+import useContract from "../useContract";
 
 export default function useUnionContract() {
   const { chainId } = useWeb3React();
