+++ hooks/useAllMemberInfo.js
@@ -2,7 +2,7 @@ import { formatUnits } from "@ethersproject/units";
 import { useWeb3React } from "@web3-react/core";
 import useSWR from "swr";
 import useCurrentToken from "./useCurrentToken";
-import useMemberContract from "./useMemberContract";
+import useMemberContract from "./contracts/useMemberContract";
 
 const getAllMemberInfo = (memberContract) => async (_, curToken, library) => {
   const signer = library.getSigner();
