+++ hooks/useStakeData.js
@@ -5,8 +5,8 @@ import useSWR from "swr";
 import { roundDown } from "util/numbers";
 import parseRes from "util/parseRes";
 import useCurrentToken from "./useCurrentToken";
-import useMemberContract from "./useMemberContract";
-import useStakingContract from "./useStakingContract";
+import useMemberContract from "./contracts/useMemberContract";
+import useStakingContract from "./contracts/useStakingContract";
 
 const getStakeData = (memberContract, stakingContract) => async (
   _,
