+++ hooks/useTrustCountData.js
@@ -2,7 +2,7 @@ import { isAddress } from "@ethersproject/address";
 import { useWeb3React } from "@web3-react/core";
 import useSWR from "swr";
 import useCurrentToken from "./useCurrentToken";
-import useMemberContract from "./useMemberContract";
+import useMemberContract from "./contracts/useMemberContract";
 
 const getTrustCount = (contract) => async (_, account, tokenAddress) => {
   let count = 0;
