+++ hooks/useTrustData.js
@@ -5,8 +5,8 @@ import LENDING_MARKET_ABI from "constants/abis/lendingMarket.json";
 import useSWR from "swr";
 import parseRes from "util/parseRes";
 import useCurrentToken from "./useCurrentToken";
-import useMarketRegistryContract from "./useMarketRegistryContract";
-import useMemberContract from "./useMemberContract";
+import useMarketRegistryContract from "./contracts/useMarketRegistryContract";
+import useMemberContract from "./contracts/useMemberContract";
 
 const getTrust = (marketRegistryContract, memberContract) => async (
   _,
