+++ views/admin.js
@@ -11,13 +11,13 @@ import INTEREST_RATE_ABI from "constants/abis/IInterestRateModel.json";
 import LENDING_MARKET_ABI from "constants/abis/lendingMarket.json";
 import { BLOCKS_PER_YEAR } from "constants/variables";
 import { useAutoCallback, useAutoEffect } from "hooks.macro";
-import useAssetContract from "hooks/useAssetContract";
-import useCompoundContract from "hooks/useCompoundContract";
+import useAssetContract from "hooks/contracts/useAssetContract";
+import useCompoundContract from "hooks/contracts/useCompoundContract";
 import useCurrentToken from "hooks/useCurrentToken";
 import useIsAdmin from "hooks/useIsAdmin";
-import useMarketRegistryContract from "hooks/useMarketRegistryContract";
-import useMemberContract from "hooks/useMemberContract";
-import useStakingContract from "hooks/useStakingContract";
+import useMarketRegistryContract from "hooks/contracts/useMarketRegistryContract";
+import useMemberContract from "hooks/contracts/useMemberContract";
+import useStakingContract from "hooks/contracts/useStakingContract";
 import { Fragment, useState } from "react";
 
 const parseRes = (res, decimals = 2) =>
